requirements.txt
dspy-ai
dspy
openai
typer[all]
python-dotenv
